fix(IMU): Remove old versions of the IMU implementation - Remove src/input/source/iio/accel_gyro_3d.rs and src/input/source/iio/bmi_imu.rs - Rename src/input/source/iio/accel_gyro_3d_new.rs and src/input/source/iio/bmi_imu_new.rs to omit the _new postfix - Update src/input/source/iio.rs with the corrected names.
